requirements.txt
pytest~=8.3.4
Faker~=33.1.0
playwright~=1.49.1
pytest-playwright==0.6.2
allure-pytest==2.13.5
allure-python-commons~=2.13.5